+++ satori/rtm/connection.py
@@ -676,9 +676,9 @@ Description
     Authorization* in the online docs.
 
 Parameters
-    * auth_delegate {AuthDelegate | RoleKeyAuthDelegate} [required] - An
+    * auth_delegate {AuthDelegate | RoleSecretAuthDelegate} [required] - An
       authentication delegate object created with
-      the `RoleKeyAuthDelegate(role, role_key)` method for
+      the `RoleSecretAuthDelegate(role, role_key)` method for
       the role-based authentication process.
     * callback {function} [required] - Function to execute after RTM
       returns a response.
@@ -689,7 +689,7 @@ Syntax
 
     secret_key = '<ROLE_SECRET_KEY>'
 
-    auth_delegate = auth.RoleKeyAuthDelegate('<ROLE>', secret_key)
+    auth_delegate = auth.RoleSecretAuthDelegate('<ROLE>', secret_key)
     auth_event = threading.Event()
 
     def auth_callback(auth_result):