+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased
+----------
+
+* Fixed RoleSecretAuthDelegate name in docstrings about authentication
+
 v1.0.0 (2017-03-07)
 -------------------
 * Initial release